+++ src/main/java/wooteco/subway/domain/Path.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.path;
+package wooteco.subway.domain;
 
 import wooteco.subway.domain.Station;
