+++ src/main/java/wooteco/subway/domain/PathFinder.java
@@ -1,8 +1,6 @@
-package wooteco.subway.domain.path;
+package wooteco.subway.domain;
 
-import wooteco.subway.domain.Station;
-
-public interface ShortestPath {
+public interface PathFinder {
     Path getPath(final Station source, final Station target, final int fare, int age);
 
     Long getExpensiveLineId(Station source, Station target);
