+++ src/main/java/wooteco/subway/infrastructure/DefaultWeightedEdgeCustom.java
@@ -1,4 +1,4 @@
-package wooteco.subway.utils;
+package wooteco.subway.infrastructure;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
